requirements.txt
aiosmtplib==3.0.1
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.8.1
backports.zoneinfo;python_version<"3.9"
bcrypt==4.1.3
beautifulsoup4==4.12.3
binaryornot==0.4.4
boto3==1.34.143
botocore==1.34.143
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
commandlines==0.4.1
cookiecutter==2.6.0
cryptography==42.0.8
decorator==4.4.2
Deprecated==1.2.13
distro==1.9.0
dnspython==2.6.1
easydict==1.13
elevenlabs==1.3.0
email_validator==2.1.1
eval_type_backport==0.2.0
exceptiongroup==1.2.1
ez_setup==0.9
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
filelock==3.14.0
filetype==1.2.0
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
fsspec==2024.6.0
future==1.0.0
gdown==5.2.0
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.12.0
hsh==1.1.0
httpcore==0.15.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.23.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsons==1.6.3
kornia==0.7.3
kornia_rs==0.1.4
lazy_loader==0.4
llvmlite==0.41.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openai==1.34.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.4
packaging==23.2
passlib==1.7.4
phonenumbers==8.12.48
Pillow==9.5.0
pkce==1.0.3
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.10.4
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMatting==1.1.12
pymongo==4.7.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
stripe==9.10.0
supertokens-python==0.22.1
sympy==1.12.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.7
tldextract==3.1.0
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transparent-background==1.3.1
twilio==7.9.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
typish==1.9.3
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.19
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
waitress==3.0.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
wget==3.2
wrapt==1.16.0
zipp==3.19.2
